requirements.txt
Flask
numpy
Pillow
tensorflow
util-linux; platform_system=="Linux"
gunicorn
